Add callback in Player.cs for successful beatmap completion
The Endless mod requires ValidForResume to be true in PlayerLoad during
regular play. Previously, OnResuming assumed the PlayerLoader screen was
returned to for the purposes of retrying the map, though this assumption
doesn't hold for Endless play. These changes allow PlayerLoader to
determine whether it was resumed for a retry or the next map in Endless
play.